Split alerts gen into dedicated files
