+++ pkg/sdk/alerts_ext.go
@@ -1,0 +1,1 @@
+package sdk
